Show password prompt in any error case in Qt Quick GUI
It might happen that an empty/incorrect password does not lead to
a crypto error. Then the data is nevertheless wrongly decrypted and
trying again with the correct password might help.

Not using PasswordFile::isEncryptionUsed() here like in the Qt Widgets
GUI for better/easier integration with Android's storage access
framework.